Solucion para el Issues #1

+++ src/controllers/productController.ts
@@ -69,7 +69,26 @@ export const productPut = async (req: Request, res: Response) => {
 
 export const productDelete = async (req: Request, res: Response) => {
     try {
-        await Product.findByIdAndDelete(req.params.id);
+        const product = await Product.findById(req.params.id, '_id category');
+        if (!product) {
+            req.log.warn(`El producto con el id ${req.params.id} no existe en la BD`);
+            return res.status(404).json({ msg: 'No existe el producto con el id: ' + req.params.id });
+        }
+        
+        const category = await Category.findOne({ name: product.category });
+        if (category) {
+            for (let i = 0; i < category.products.length!; i++) { 
+                if (category.products[i].equals(product._id)) {
+                    category.products.splice(i, 1);
+                    break;
+                }
+            }
+        }
+        
+        await Promise.all([
+            category?.save(),
+            Product.findByIdAndDelete(req.params.id)
+        ]);
         res.json({ msg: 'Producto eliminado' });
         req.log.info('Elimino el producto con el id: ' + req.params.id);
 
@@ -81,11 +100,6 @@ export const productDelete = async (req: Request, res: Response) => {
 
 export const updatePicture = async (req: Request, res: Response) => {
     try {
-        if (!req.files || Object.keys(req.files).length === 0 || !req.files.file) {
-            req.log.warn('No hay archivo para subir');
-            return res.status(400).json({ msg: 'No hay archivo para subir' });
-        }
-
         const product = await Product.findById(req.params.id);
         if (!product) {
             req.log.warn(`El producto con el id ${req.params.id} no existe en la BD`);
@@ -96,7 +110,7 @@ export const updatePicture = async (req: Request, res: Response) => {
             deleteFile(product.picture);
         }
 
-        product.picture = await upload(req.files.file) || '';
+        product.picture = await upload(req.files!.file) || '';
         await product.save();
         res.json(product);
         req.log.info('Actualizo la imagen del producto: ' + product._id);
